Prepared to work easier with revision and speculative metadata for files introduced in apiv2
